all methods of DaraLinker integrated into NumericInformationExtractor, MetaDataExtractor and ReferenceResolver; NumericInformationExtractor no longer duplicates code in RegexUtils; cleaned RegexUtils (#13); all tests from DaraLinkerTest integrated into MetaDataExtractorTest and RegexUtilsTest
